+++ router-tests/structured_logging_test.go
@@ -6,6 +6,7 @@ import (
 	"github.com/wundergraph/cosmo/router/core"
 	"github.com/wundergraph/cosmo/router/pkg/config"
 	"github.com/wundergraph/cosmo/router/pkg/logging"
+	"github.com/wundergraph/cosmo/router/pkg/trace/tracetest"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
 	"math"
@@ -88,7 +89,7 @@ func TestAccessLogsFileOutput(t *testing.T) {
 			require.NoError(t, os.RemoveAll(fp))
 		})
 
-		logger := logging.NewZapAccessLogger(f, false)
+		logger := logging.NewZapAccessLogger(f, false, false)
 		require.NoError(t, err)
 
 		testenv.Run(t, &testenv.Config{
@@ -162,6 +163,42 @@ func TestAccessLogs(t *testing.T) {
 			}
 			additionalExpectedKeys := []string{
 				"user_agent", "latency", "config_version", "request_id", "pid", "hostname",
+			}
+			checkValues(t, requestContext, expectedValues, additionalExpectedKeys)
+		})
+	})
+
+	t.Run("Simple with tracing enabled", func(t *testing.T) {
+		t.Parallel()
+
+		exporter := tracetest.NewInMemoryExporter(t)
+
+		testenv.Run(t, &testenv.Config{
+			LogObservation: testenv.LogObservationConfig{
+				Enabled:  true,
+				LogLevel: zapcore.InfoLevel,
+			},
+			TraceExporter: exporter,
+		}, func(t *testing.T, xEnv *testenv.Environment) {
+			res := xEnv.MakeGraphQLRequestOK(testenv.GraphQLRequest{
+				Query: `{ employees { id } }`,
+			})
+			require.JSONEq(t, employeesIDData, res.Body)
+			logEntries := xEnv.Observer().All()
+			require.Len(t, logEntries, 12)
+			requestLog := xEnv.Observer().FilterMessage("/graphql")
+			require.Equal(t, requestLog.Len(), 1)
+			requestContext := requestLog.All()[0].ContextMap()
+			expectedValues := map[string]interface{}{
+				"log_type": "request",
+				"status":   int64(200),
+				"method":   "POST",
+				"path":     "/graphql",
+				"query":    "", // http query is empty
+				"ip":       "[REDACTED]",
+			}
+			additionalExpectedKeys := []string{
+				"user_agent", "latency", "config_version", "request_id", "pid", "hostname", "trace_id",
 			}
 			checkValues(t, requestContext, expectedValues, additionalExpectedKeys)
 		})
@@ -326,7 +363,7 @@ func TestAccessLogs(t *testing.T) {
 		})
 	})
 
-	t.Run("Log when operation parsing fails", func(t *testing.T) {
+	t.Run("Log as much information possible when operation parsing fails", func(t *testing.T) {
 		t.Parallel()
 
 		testenv.Run(t, &testenv.Config{
@@ -336,6 +373,12 @@ func TestAccessLogs(t *testing.T) {
 					Default: "",
 					ValueFrom: &config.CustomDynamicAttribute{
 						RequestHeader: "service-name",
+					},
+				},
+				{
+					Key: "error_message",
+					ValueFrom: &config.CustomDynamicAttribute{
+						ContextField: core.ContextFieldResponseErrorMessage,
 					},
 				},
 				{
@@ -407,14 +450,15 @@ func TestAccessLogs(t *testing.T) {
 			require.Equal(t, requestLog.Len(), 1)
 			requestContext := requestLog.All()[0].ContextMap()
 			expectedValues := map[string]interface{}{
-				"log_type":     "request",
-				"status":       int64(200),
-				"method":       "POST",
-				"path":         "/graphql",
-				"query":        "", // http query is empty
-				"ip":           "[REDACTED]",
-				"user_agent":   "Go-http-client/1.1",
-				"service_name": "service-name", // From header
+				"log_type":      "request",
+				"status":        int64(200),
+				"method":        "POST",
+				"path":          "/graphql",
+				"query":         "", // http query is empty
+				"ip":            "[REDACTED]",
+				"user_agent":    "Go-http-client/1.1",
+				"service_name":  "service-name", // From header
+				"error_message": "unexpected token - got: EOF want one of: [RBRACE IDENT SPREAD]",
 			}
 			additionalExpectedKeys := []string{
 				"latency",
@@ -429,7 +473,7 @@ func TestAccessLogs(t *testing.T) {
 		})
 	})
 
-	t.Run("Log when operation normalization fails", func(t *testing.T) {
+	t.Run("Log as much information possible when operation normalization fails", func(t *testing.T) {
 		t.Parallel()
 
 		testenv.Run(t, &testenv.Config{
@@ -439,6 +483,12 @@ func TestAccessLogs(t *testing.T) {
 					Default: "",
 					ValueFrom: &config.CustomDynamicAttribute{
 						RequestHeader: "service-name",
+					},
+				},
+				{
+					Key: "error_message",
+					ValueFrom: &config.CustomDynamicAttribute{
+						ContextField: core.ContextFieldResponseErrorMessage,
 					},
 				},
 				{
@@ -520,6 +570,7 @@ func TestAccessLogs(t *testing.T) {
 				"service_name":   "service-name", // From header
 				"operation_type": "query",        // From context
 				"operation_name": "employees",    // From context
+				"error_message":  "field: notExists not defined on type: Query",
 			}
 			additionalExpectedKeys := []string{
 				"latency",
@@ -535,7 +586,7 @@ func TestAccessLogs(t *testing.T) {
 		})
 	})
 
-	t.Run("Log when panic occurs on execution / error panic", func(t *testing.T) {
+	t.Run("Log as much information possible on execution / error panic", func(t *testing.T) {
 		t.Parallel()
 
 		testenv.Run(t, &testenv.Config{
@@ -545,6 +596,12 @@ func TestAccessLogs(t *testing.T) {
 					Default: "",
 					ValueFrom: &config.CustomDynamicAttribute{
 						RequestHeader: "service-name",
+					},
+				},
+				{
+					Key: "error_message",
+					ValueFrom: &config.CustomDynamicAttribute{
+						ContextField: core.ContextFieldResponseErrorMessage,
 					},
 				},
 				{
@@ -637,6 +694,7 @@ func TestAccessLogs(t *testing.T) {
 				"operation_hash": "14226210703439426856", // From context
 				"operation_name": "employees",            // From context
 				"operation_type": "query",                // From context
+				"error_message":  "implement me",
 			}
 			additionalExpectedKeys := []string{
 				"latency",
@@ -655,7 +713,7 @@ func TestAccessLogs(t *testing.T) {
 		})
 	})
 
-	t.Run("Log when panic occurs on execution / string panic", func(t *testing.T) {
+	t.Run("Log as much information possible on execution / string panic", func(t *testing.T) {
 		t.Parallel()
 
 		testenv.Run(t, &testenv.Config{
@@ -665,6 +723,12 @@ func TestAccessLogs(t *testing.T) {
 					Default: "",
 					ValueFrom: &config.CustomDynamicAttribute{
 						RequestHeader: "service-name",
+					},
+				},
+				{
+					Key: "error_message",
+					ValueFrom: &config.CustomDynamicAttribute{
+						ContextField: core.ContextFieldResponseErrorMessage,
 					},
 				},
 				{
@@ -758,6 +822,7 @@ func TestAccessLogs(t *testing.T) {
 				"operation_hash": "14226210703439426856", // From context
 				"operation_name": "employees",            // From context
 				"operation_type": "query",                // From context
+				"error_message":  "implement me",
 			}
 			additionalExpectedKeys := []string{
 				"latency",