+++ cli/CHANGELOG.md
@@ -4,6 +4,14 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.68.3](https://github.com/wundergraph/cosmo/compare/wgc@0.68.2...wgc@0.68.3) (2024-10-16)
+
+**Note:** Version bump only for package wgc
+
+## [0.68.2](https://github.com/wundergraph/cosmo/compare/wgc@0.68.1...wgc@0.68.2) (2024-10-15)
+
+**Note:** Version bump only for package wgc
+
 ## [0.68.1](https://github.com/wundergraph/cosmo/compare/wgc@0.68.0...wgc@0.68.1) (2024-10-14)
 
 **Note:** Version bump only for package wgc